requirements.txt
cycler==0.10.0
decorator==4.4.0
dlib==19.17.0
gdbn==0.1
gnumpy==0.2
imageio==2.5.0
imutils==0.5.2
joblib==0.13.2
kiwisolver==1.0.1
Lasagne==0.1
matplotlib==3.0.3
networkx==2.2
nolearn==0.6.0
numpy==1.16.2
opencv-python==4.1.0.25
pandas==0.24.2
Pillow==6.0.0
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2018.9
PyWavelets==1.0.2
scikit-image==0.15.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
sklearn==0.0
tabulate==0.8.3
Theano==1.0.4
torch==1.0.1.post2
